requirements.txt
altgraph==0.17.4
asarPy==1.0.1
packaging==24.2
pillow==11.1.0
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
setuptools==75.8.2
